Added pagination and sorting to post apis

+++ quillpost/post/blog-post-service/src/main/java/com/quillpost/blogpostservice/payloads/CategoryDTO.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @Setter
 public class CategoryDTO {
     
-    private Long categoryID;
+    private Integer categoryID;
     private String categoryName;
 }
